+++ packages/io/fileio/current/tests/testfs.c
@@ -1166,7 +1166,8 @@ void testfs_dump(void)
         {
             // Regular file
             int j;
-            diag_printf("f %8s %4d |",nd->name,nd->status.st_size);
+            diag_printf("f %8s %4ld |",nd->name,
+                        (unsigned long)nd->status.st_size);
             for( j = 0; j < TESTFS_FILEBLOCKS; j++ )
             {
                 testfs_block *b = nd->u.file.data[j];
@@ -1174,7 +1175,8 @@ void testfs_dump(void)
                 {
                     if( j > 0 && (j%4) == 0 )
                         diag_printf(indent);
-                    diag_printf(" %3d[%3d,%3d]",b-block,b->pos,b->size);
+                    diag_printf(" %3d[%3ld,%3d]",b-block,
+                                (unsigned long)b->pos,b->size);
                     if( b->u.file != nd )
                     {
                         errors++;
@@ -1226,7 +1228,7 @@ void testfs_dump(void)
         {
             int j;
             testfs_node *nd = b->u.file;
-            diag_printf(" %3d %3d %d[%7s]",b->pos,b->size,nd-node,nd->name);
+            diag_printf(" %3ld %3d %d[%7s]",b->pos,b->size,nd-node,nd->name);
             for( j = 0; j < TESTFS_FILEBLOCKS; j++ )
             {
                 if( nd->u.file.data[j] == b )
